ReportTableModelMapper: Extract model class into a separate file
This separates the mapper from the model and make it easier to read
the code of the mapper. This prepares for the extraction of several
functions inside the mapper which are then easier to find since
they are not intermixed with a bunch of data classes.
